feat: Prevent deleting or editing by accessing manually URLs

+++ views.py
@@ -1,4 +1,4 @@
-from database_setup import Genre, Album, Song
+from database_setup import Genre, Album, Song, User
 from addcollection import session
 from flask import (Flask, render_template, redirect, url_for, request,
                    flash, jsonify, make_response)
@@ -17,11 +17,12 @@ CLIENT_ID = json.loads(
 
 app = Flask(__name__)
 
-# Anti Foregery token and login------------------------
+# Anti Forgery token and login------------------------
 
 
 @app.route('/login/')
 def showLogin():
+    """Create anti forgery token and render login page."""
     # generate random token
     state = ''.join(
         random.choice(
@@ -35,6 +36,7 @@ def showLogin():
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
+    """Manage google sign in via OAuth API."""
     # Anti forgery token check
     if request.args.get('state') != login_session['state']:
         response = make_response(json.dumps('Invalid state parameter'), 401)
@@ -60,7 +62,7 @@ def gconnect():
 
     # Check if acess token in credentials object is valid
     access_token = credentials.access_token
-    url = 'https://www.googleapis.com/oauth2/v1/' +
+    url = 'https://www.googleapis.com/oauth2/v1/'\
           'tokeninfo?access_token={}'.format(access_token)
     result = json.loads(httplib2.Http().request(url, 'GET')[1])
     # if error in access token, abort
@@ -102,6 +104,11 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+    # Retrieve user_id if user exists in DB, or create a new one if it doesn't
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        login_session['user_id'] = createUser(login_session)
+
     output = """
     <h1>
         Welcome, {}
@@ -118,9 +125,9 @@ def gconnect():
     return output
 
 
-# revoke access token and reset login_session
 @app.route('/gdisconnect/')
 def gdisconnect():
+    """Disconnect user from application."""
     # Only disconnect a connected user
     access_token = login_session['access_token']
     if access_token is None:
@@ -129,7 +136,7 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     # revoke current user
-    url = 'https://accounts.google.com/o/oauth2/' +
+    url = 'https://accounts.google.com/o/oauth2/'\
           'revoke?token={}'.format(access_token)
     result = httplib2.Http().request(url, 'GET')[0]
     # if request successful reset session
@@ -153,18 +160,21 @@ def gdisconnect():
 # API Endpoints ---------------------------------
 @app.route('/genres/json/')
 def genresJSON():
+    """Handle genres JSON API Endpoint."""
     genres = session.query(Genre).all()
     return jsonify([genre.serialize for genre in genres])
 
 
 @app.route('/genre/<int:genre_id>/json/')
 def releasesJSON(genre_id):
+    """Handle specific genre JSON API Endpoint."""
     releases = session.query(Album).filter_by(genre_id=genre_id).all()
     return jsonify([release.serialize for release in releases])
 
 
 @app.route('/release/<int:release_id>/json/')
 def songsJSON(release_id):
+    """Handle songs JSON API Endpoint."""
     songs = session.query(Song).filter_by(release_id=release_id).all()
     return jsonify([song.serialize for song in songs])
 
@@ -172,26 +182,34 @@ def songsJSON(release_id):
 @app.route('/')
 @app.route('/genres/')
 def showGenres():
+    """Render template displaying all genres."""
     # check if user logged in:
-    user = login_session['username']
+    if 'username' not in login_session:
+        logged = False
+    else:
+        logged = True
     genres_count = (session
                     .query(Genre.id, Genre.name, func.count(Album.genre_id))
                     .outerjoin(Album)
                     .group_by(Genre)
                     .order_by(Genre.name)
                     .all())
-    return render_template('genres.html', genres_count=genres_count, user=user)
+    return render_template('genres.html',
+                           genres_count=genres_count,
+                           logged=logged)
 
 
 # -------------------------------------------------- GENRES ROUTES
 @app.route('/genre/new/', methods=['GET', 'POST'])
 def newGenre():
+    """Render template to add new genre and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
         if not (session
                 .query(Genre).filter_by(name=request.form['name']).first()):
-            session.add(Genre(name=request.form['name']))
+            session.add(Genre(name=request.form['name'],
+                              user_id=login_session['user_id']))
             session.commit()
             flash('New Genre added!')
             return redirect(url_for('showGenres'))
@@ -204,9 +222,20 @@ def newGenre():
 
 @app.route('/genre/<int:genre_id>/edit/', methods=['GET', 'POST'])
 def editGenre(genre_id):
+    """Render template to edit genre and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     genre = session.query(Genre).filter_by(id=genre_id).one()
+    if login_session['user_id'] != genre.user_id:
+        return """
+            <script>
+                function myFunction() {
+                    alert('You are not authorized to edit menu items to
+                    this restaurant. Please create your own restaurant in
+                    order to edit items.');
+                }
+            </script><body onload='myFunction()''>
+            """
     if request.method == 'POST':
         if request.form['name']:
             genre.name = request.form['name']
@@ -219,9 +248,20 @@ def editGenre(genre_id):
 
 @app.route('/genre/<int:genre_id>/delete/', methods=['GET', 'POST'])
 def deleteGenre(genre_id):
+    """Render template to delete genre and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     genre = session.query(Genre).filter_by(id=genre_id).one()
+    if login_session['user_id'] != genre.user_id:
+        return """
+            <script>
+                function myFunction() {
+                    alert('You are not authorized to edit menu items to
+                    this restaurant. Please create your own restaurant in
+                    order to edit items.');
+                }
+            </script><body onload='myFunction()''>
+            """
     if request.method == 'POST':
         session.delete(genre)
         session.commit()
@@ -234,15 +274,25 @@ def deleteGenre(genre_id):
 # -------------------------------------------------- RELEASES ROUTES
 @app.route('/genre/<int:genre_id>/')
 def showReleases(genre_id):
+    """Render template to show all releases of one genre."""
+    # check if user logged
+    if 'username' not in login_session:
+        logged = False
+    else:
+        logged = True
     genre = session.query(Genre).filter_by(id=genre_id).one()
     releases = (session
                 .query(Album)
                 .filter_by(genre_id=genre_id).order_by(Album.title).all())
-    return render_template('releases.html', genre=genre, releases=releases)
+    return render_template('releases.html',
+                           genre=genre,
+                           releases=releases,
+                           logged=logged)
 
 
 @app.route('/genre/<int:genre_id>/new/', methods=['GET', 'POST'])
 def newRelease(genre_id):
+    """Render template to add new release and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     genre = session.query(Genre).filter_by(id=genre_id).one()
@@ -252,7 +302,8 @@ def newRelease(genre_id):
                         label=request.form['label'],
                         released=request.form['released'],
                         image=request.form['image'],
-                        genre_id=genre.id)
+                        genre_id=genre.id,
+                        user_id=genre.user_id)
         session.add(release)
         session.commit()
         flash('New release added!')
@@ -264,10 +315,21 @@ def newRelease(genre_id):
 @app.route('/genre/<int:genre_id>/<int:release_id>/edit/',
            methods=['GET', 'POST'])
 def editRelease(genre_id, release_id):
+    """Render template to edit a release and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     genre = session.query(Genre).filter_by(id=genre_id).one()
     release = session.query(Album).filter_by(id=release_id).one()
+    if login_session['user_id'] != release.user_id:
+        return """
+            <script>
+                function myFunction() {
+                    alert('You are not authorized to edit menu items to
+                    this restaurant. Please create your own restaurant in
+                    order to edit items.');
+                }
+            </script><body onload='myFunction()''>
+            """
     if request.method == 'POST':
         if request.form['title']:
             release.title = request.form['title']
@@ -289,10 +351,21 @@ def editRelease(genre_id, release_id):
 @app.route('/genre/<int:genre_id>/<int:release_id>/delete/',
            methods=['GET', 'POST'])
 def deleteRelease(genre_id, release_id):
+    """Render template to delete a release and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     genre = session.query(Genre).filter_by(id=genre_id).one()
     release = session.query(Album).filter_by(id=release_id).one()
+    if login_session['user_id'] != release.user_id:
+        return """
+            <script>
+                function myFunction() {
+                    alert('You are not authorized to edit menu items to
+                    this restaurant. Please create your own restaurant in
+                    order to edit items.');
+                }
+            </script><body onload='myFunction()''>
+            """
     if request.method == 'POST':
         session.delete(release)
         session.commit()
@@ -306,22 +379,33 @@ def deleteRelease(genre_id, release_id):
 # -------------------------------------------------- SONGS ROUTES
 @app.route('/release/<int:release_id>/')
 def showSongs(release_id):
+    """Render template to show all songs."""
+    # check if user logged
+    if 'username' not in login_session:
+        logged = False
+    else:
+        logged = True
     release = session.query(Album).filter_by(id=release_id).one()
     songs = (session
              .query(Song)
              .filter_by(release_id=release_id).order_by(Song.position).all())
-    return render_template('songs.html', release=release, songs=songs)
+    return render_template('songs.html',
+                           release=release,
+                           songs=songs,
+                           logged=logged)
 
 
 @app.route('/release/<int:release_id>/new/', methods=['GET', 'POST'])
 def newSong(release_id):
+    """Render template to add song to a release and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     release = session.query(Album).filter_by(id=release_id).one()
     if request.method == 'POST':
         song = Song(title=request.form['title'],
                     position=request.form['position'],
-                    release_id=release.id)
+                    release_id=release.id,
+                    user_id=release.user_id)
         session.add(song)
         session.commit()
         flash('New song added!')
@@ -333,10 +417,21 @@ def newSong(release_id):
 @app.route('/release/<int:release_id>/<int:song_id>/edit/',
            methods=['GET', 'POST'])
 def editSong(song_id, release_id):
+    """Render template to edit a song and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     song = session.query(Song).filter_by(id=song_id).one()
     release = session.query(Album).filter_by(id=release_id).one()
+    if login_session['user_id'] != song.user_id:
+        return """
+            <script>
+                function myFunction() {
+                    alert('You are not authorized to edit menu items to
+                    this restaurant. Please create your own restaurant in
+                    order to edit items.');
+                }
+            </script><body onload='myFunction()''>
+            """
     if request.method == 'POST':
         if request.form['title']:
             song.title = request.form['title']
@@ -352,10 +447,21 @@ def editSong(song_id, release_id):
 @app.route('/release/<int:release_id>/<int:song_id>/delete/',
            methods=['GET', 'POST'])
 def deleteSong(song_id, release_id):
+    """Render template to delete a song and handle post request."""
     if 'username' not in login_session:
         return redirect('/login')
     song = session.query(Song).filter_by(id=song_id).one()
     release = session.query(Album).filter_by(id=release_id).one()
+    if login_session['user_id'] != song.user_id:
+        return """
+            <script>
+                function myFunction() {
+                    alert('You are not authorized to edit menu items to
+                    this restaurant. Please create your own restaurant in
+                    order to edit items.');
+                }
+            </script><body onload='myFunction()''>
+            """
     if request.method == 'POST':
         session.delete(song)
         session.commit()
@@ -363,6 +469,37 @@ def deleteSong(song_id, release_id):
         return redirect(url_for('showSongs', release_id=release_id))
     if request.method == 'GET':
         return render_template('deletesong.html', song=song, release=release)
+
+
+def createUser(login_session):
+    """Create new user in DB."""
+    newUser = User(name=login_session['username'],
+                   email=login_session['email'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    """Get user info from DB.
+
+    arg: user ID integer
+    """
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    """Get user ID from DB.
+
+    arg: user's email string
+    """
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except Exception:
+        return None
 
 
 if __name__ == '__main__':
